requirements.txt
asttokens==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
executing==1.1.1
future==0.18.2
gkeepapi==0.14.2
gpsoauth==1.0.2
icecream==2.1.3
idna==3.4
pycryptodomex==3.15.0
Pygments==2.13.0
python-decouple==3.6
requests==2.28.1
six==1.16.0
urllib3==1.26.12
